backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-Migrate==4.0.5
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
